requirements.txt
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.1.1
ecdsa==0.14.1
Faker==4.14.0
fastapi==0.61.1
h11==0.11.0
idna==2.10
iniconfig==1.1.1
packaging==20.4
passlib==1.7.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.1.1
python-dateutil==2.8.1
python-jose==3.2.0
python-multipart==0.0.5
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.10
uvicorn==0.12.1
